+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add Labels and Annotations to ServiceMonitor
 - Allow to expose Exporter in HTTP with secured Deployments
 - Change rotation by annotation order (coordinator before dbserver)
+- Fix NodeAffinity propagation
 
 ## [1.0.4](https://github.com/arangodb/kube-arangodb/tree/1.0.4) (2020-07-28)
 - Add Encryption Key rotation feature for ArangoDB EE 3.7+